Rewrite tour.js
This was inspired by ionic-tour plugin.
The Scrolling is calculated by first adding the new tourtip and checking
then if it is out of screen. Then the tourtip and the screen are scrolled.
Also are all position and offsets now retrieved by the $ionicPosition
injector and not by jQuery.
Furthermore an overlaying div stops the user from interacting with the app
while the tour is running.

Successfully tested on sony z10.